+++ packages/plugin-babel/src/index.ts
@@ -1,0 +1,1 @@
+export {babel} from './plugin-babel';