+++ packages/plugin-javascript/src/babel-preset/index.ts
@@ -1,2 +1,0 @@
-export {default} from './babel-preset';
-export type {Module, Polyfill, Target, Options} from './babel-preset';