+++ packages/plugin-javascript/src/index.ts
@@ -1,7 +1,0 @@
-export {javascript, babelPlugins, babelPresets} from './plugin-javascript';
-export {
-  updateSewingKitBabelPreset,
-  updateBabelPlugin,
-  updateBabelPreset,
-} from './utilities';
-export type {BabelConfig, BabelHooks} from './types';