+++ packages/plugin-package-build/CHANGELOG.md
@@ -7,6 +7,11 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Breaking Change
+
+- Load babel config from `plugin-babel` instead of `plugin-javascript`. [[#218](https://github.com/Shopify/sewing-kit-next/pull/218)]
+- The hardcoded `configFile: false` option passed to `@rollup/babel` has been removed. This means that Babel can look at `babel.config.js` files for config. If you wish to disable this, set `configFile: false` explicitly when configuring the sewing-kit's `babel()` plugin. [[#218](https://github.com/Shopify/sewing-kit-next/pull/218)]
+
 ### Changed
 
 - Removed js output from `build/ts` folder. [[#217](https://github.com/Shopify/sewing-kit-next/pull/217)]
@@ -27,7 +32,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Added package metadata and limit files included in packages. Ensure `*.tsbuildinfo` files are absent to greatly shink package size. [[#177](https://github.com/Shopify/sewing-kit-next/pull/177)]
 - Remove unused dependency on `@babel/core` [[#177](https://github.com/Shopify/sewing-kit-next/pull/177)]
-- Remove dependencies on `@sewing-kit/plugin-javascript` and `@sewing-kit/plugin-typescript`. `plugin-package-build` is now responsible for writing js/mjs/esnext/d.ts entry files.
+- Remove dependencies on `@sewing-kit/plugin-babel` and `@sewing-kit/plugin-typescript`. `plugin-package-build` is now responsible for writing js/mjs/esnext/d.ts entry files.
 
 ## 0.2.4 - 2021-06-18